task 6 square generator class in separate module
