+++ main.py
@@ -1,15 +1,4 @@
-import math
-
-
-class SquareGenerator:
-    @staticmethod
-    def generate_squares(first, last):
-        if last < first:
-            first, last = last, first
-            print("first was > then second so switched")
-        squares = [(x, math.sqrt(x)) for x in range(first, last + 1)]
-        return squares
-
+from square_generator import SquareGenerator
 
 if __name__ == "__main__":
     square_generator = SquareGenerator()
